# contrato/openapi.json
{
  "components": {
    "schemas": {
...
      "CambioEntrada": {
        "additionalProperties": false,
        "description": "La misma peticion y la lista que el lector acepta, que tiene que ser la\npropuesta (`RF-48`, `RF-50`).",
        "properties": {
          "capitulos_propuestos": {
            "items": {
              "type": "string"
            },
            "title": "Capitulos Propuestos",
            "type": "array"
          },
          "clase": {
            "$ref": "#/components/schemas/ClaseDePeticion"
          },
          "enunciado_nuevo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enunciado Nuevo"
          },
          "hecho": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hecho"
          },
          "nombre_nuevo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Nuevo"
          },
          "personaje": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaje"
          },
          "texto": {
            "title": "Texto",
            "type": "string"
          },
          "version_de_partida": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version De Partida"
          }
        },
        "required": [
          "clase",
          "texto",
          "capitulos_propuestos"
        ],
        "title": "CambioEntrada",
        "type": "object"
      },
...
      "EscenaLeida": {
        "description": "Una escena para leer: la del indice mas su texto. **Nunca** el texto sin estado.",
        "properties": {
          "borrador": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BorradorElegido"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nulo si la escena no tiene ningun borrador (p. ej. planificada)"
          },
          "capitulo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "El id del Capitulo que la contiene; nunca el de la obra. Nulo solo en escaletas anteriores a SPEC-21, que no lo guardaban",
            "title": "Capitulo"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoDeEscena"
          },
          "hallazgos_abiertos": {
            "items": {
              "$ref": "#/components/schemas/HallazgoAbierto"
            },
            "title": "Hallazgos Abiertos",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "personajes_presentes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Escena.personajes_presentes: los que la pagina ofrece para renombrar. Nulo si la escena no los declara: no declarado, que no es nadie",
            "title": "Personajes Presentes"
          },
          "se_acepto_rindiendose": {
            "description": "Resuelto en el backend: estado == aceptada_por_rendicion (RF-40)",
            "title": "Se Acepto Rindiendose",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "capitulo",
          "estado",
          "se_acepto_rindiendose",
          "hallazgos_abiertos",
          "borrador",
          "personajes_presentes"
        ],
        "title": "EscenaLeida",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VersionesSalida": {
        "additionalProperties": false,
        "properties": {
          "obra": {
            "title": "Obra",
            "type": "string"
          },
          "versiones": {
            "items": {
              "$ref": "#/components/schemas/VersionSalida"
            },
            "title": "Versiones",
            "type": "array"
          },
          "vigente": {
            "description": "El numero de la version vigente: la ultima publicada, o la 1 si no hay ninguna (`F-121`, `F-150`)",
            "title": "Vigente",
            "type": "integer"
          }
        },
        "required": [
          "obra",
          "versiones",
          "vigente"
        ],
        "title": "VersionesSalida",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Harness de novelas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/capitulos/{id_capitulo}/cerrar": {
      "post": {
        "description": "La segunda puerta con firma humana.\n\nDevuelve los `menor` que se dejan pasar: si no se enseñaran, `mayor` y\n`menor` volverian a producir el mismo comportamiento.",
        "operationId": "cerrar_capitulo_capitulos__id_capitulo__cerrar_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_capitulo",
            "required": true,
            "schema": {
              "title": "Id Capitulo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cerrar Capitulo",
        "tags": [
          "ciclo"
        ]
      }
    },
...
    "/escenas/{id_escena}": {
      "get": {
        "description": "Una escena con su estado, sus hallazgos abiertos y su texto elegido (`RF-39`).",
        "operationId": "escena_escenas__id_escena__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_escena",
            "required": true,
            "schema": {
              "title": "Id Escena",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscenaLeida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Escena",
        "tags": [
          "lectura"
        ]
      }
    },
    "/escenas/{id_escena}/aceptar": {
      "post": {
        "description": "La dispara una persona. El worker no tiene ninguna ruta que llegue aqui.\n\n`rindiendose` deja la escena en `aceptada_por_rendicion`, que **no es el\nmismo hecho** que `aceptada`: su delta entra igual al canon y quien lea el\nmanuscrito despues necesita saber cuales fueron.",
        "operationId": "aceptar_escenas__id_escena__aceptar_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_escena",
            "required": true,
            "schema": {
              "title": "Id Escena",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "rindiendose",
            "required": false,
            "schema": {
              "default": false,
              "title": "Rindiendose",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aceptar",
        "tags": [
          "ciclo"
        ]
      }
    },
    "/escenas/{id_escena}/generar": {
      "post": {
        "description": "`202` y un identificador. Nunca el texto, y nunca sincrono.",
        "operationId": "generar_escenas__id_escena__generar_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_escena",
            "required": true,
            "schema": {
              "title": "Id Escena",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generar",
        "tags": [
          "ciclo"
        ]
      }
    },
...
    "/escenas/{id_escena}/rechazar": {
      "post": {
        "operationId": "rechazar_escenas__id_escena__rechazar_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_escena",
            "required": true,
            "schema": {
              "title": "Id Escena",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "motivo",
            "required": true,
            "schema": {
              "title": "Motivo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rechazar",
        "tags": [
          "ciclo"
        ]
      }
    },
...
    "/obras/{id_obra}/cambios": {
      "post": {
        "description": "`202` con el identificador de trabajo. `409` si no hay salida elegida -hoy,\nsiempre: falta la medida-, si la lista no es la propuesta o si `C-4` no la admite.\nCon `409` **no se guarda ni se encola nada**.",
        "operationId": "cambios_obras__id_obra__cambios_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cambios",
        "tags": [
          "ciclo"
        ]
      }
    },
    "/obras/{id_obra}/cambios/propuesta": {
      "post": {
        "description": "Los capitulos que se van a tocar, **antes** de tocarlos, con la promesa y su\npunto ciego (`RF-51`, `RF-55`). Sincrona y sin modelo. `409` si `C-4` no la admite.",
        "operationId": "propuesta_obras__id_obra__cambios_propuesta_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropuestaSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propuesta",
        "tags": [
          "ciclo"
        ]
      }
    },
    "/obras/{id_obra}/entregar": {
      "post": {
        "description": "`SPEC-25` `RF-21`: borra la entrevista y conserva la novela, sus vetadas y\nsus hechos. Devuelve cuantas filas se borraron, nunca que contenian.",
        "operationId": "entregar_obras__id_obra__entregar_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entregar",
        "tags": [
          "ciclo"
        ]
      }
    },
...
    "/obras/{id_obra}/versiones": {
      "get": {
        "description": "Las versiones con su numero, su anterior, su commit, cuando se crearon y su\npeticion (`RF-53`), y cual es la vigente: la web no la deduce (`F-150`).",
        "operationId": "versiones_obras__id_obra__versiones_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionesSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Versiones",
        "tags": [
          "ciclo"
        ]
      }
    },
    "/obras/{id_obra}/versiones/{numero}": {
      "get": {
        "description": "Los capitulos en orden: si es compartido con la anterior, su estado, y por\nescena su `estado_de_escena` y su `estado_de_verificacion` (`RF-52`..`RF-54`).",
        "operationId": "version_obras__id_obra__versiones__numero__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "numero",
            "required": true,
            "schema": {
              "title": "Numero",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionDetalleSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Version",
        "tags": [
          "ciclo"
        ]
      }
    },
    "/obras/{id_obra}/versiones/{numero}/capitulos/{id_capitulo}": {
      "get": {
        "description": "Un capitulo de una version, entero. `404` si no es de esa version (`RF-53`).",
        "operationId": "capitulo_de_version_obras__id_obra__versiones__numero__capitulos__id_capitulo__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "numero",
            "required": true,
            "schema": {
              "title": "Numero",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id_capitulo",
            "required": true,
            "schema": {
              "title": "Id Capitulo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapituloLeidoDeVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capitulo De Version",
        "tags": [
          "ciclo",
          "lectura"
        ]
      }
    },
    "/obras/{id_obra}/versiones/{numero}/indice": {
      "get": {
        "description": "El indice de una version: cada capitulo con su marca de cambio y cada escena con\nsu estado, sus hallazgos y su `estado_de_verificacion` en esa version (`RF-52`..`RF-54`).",
        "operationId": "indice_de_version_obras__id_obra__versiones__numero__indice_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "numero",
            "required": true,
            "schema": {
              "title": "Numero",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndiceDeVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Indice De Version",
        "tags": [
          "ciclo",
          "lectura"
        ]
      }
    },
...
    "/trabajos/{id_trabajo}": {
      "get": {
        "description": "El estado de un trabajo encolado: su resultado, o el motivo si fallo.\n\nEs la otra mitad de todos los `202`: sin esto el cliente sabe que se encolo\nalgo y no tiene forma de saber que paso.",
        "operationId": "consultar_trabajo_trabajos__id_trabajo__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_trabajo",
            "required": true,
            "schema": {
              "title": "Id Trabajo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajoSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consultar Trabajo",
        "tags": [
          "ciclo"
        ]
      }
    }
  }
}